+++ src/test/java/com/picimako/justkitting/ContentRootsProjectDescriptor.java
@@ -10,6 +10,7 @@ import com.intellij.openapi.roots.ModifiableRootModel;
 import com.intellij.testFramework.fixtures.DefaultLightProjectDescriptor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.jps.model.java.JavaResourceRootType;
+import org.jetbrains.jps.model.java.JavaSourceRootType;
 
 /**
  * Project descriptor that loads main/ and test/ sources and resources folders.
@@ -28,7 +29,7 @@ public class ContentRootsProjectDescriptor extends DefaultLightProjectDescriptor
         contentEntry.clearSourceFolders();
 
         String contentEntryUrl = contentEntry.getUrl();
-//        contentEntry.addSourceFolder(contentEntryUrl + "/main/java", JavaSourceRootType.SOURCE);
+        contentEntry.addSourceFolder(contentEntryUrl + "/main/java", JavaSourceRootType.SOURCE);
         contentEntry.addSourceFolder(contentEntryUrl + "/main/resources", JavaResourceRootType.RESOURCE);
 //        contentEntry.addSourceFolder(contentEntryUrl + "/test/java", JavaSourceRootType.TEST_SOURCE);
 //        contentEntry.addSourceFolder(contentEntryUrl + "/test/resources", JavaResourceRootType.TEST_RESOURCE);